requirements.txt
streamlit==1.28.1
google-generativeai==0.3.0
pandas==2.1.1
scikit-learn==1.3.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
python-dotenv==1.0.0
transformers==4.35.0
torch>=1.9.0
sentencepiece==0.1.99 